requirements.txt
# Data collection & storage
requests==2.32.3
aiohttp==3.10.5

# Database
SQLAlchemy==2.0.36

# Data processing
pandas==2.2.3
numpy==2.1.3
pyarrow==18.0.0

# ML
lightgbm==4.5.0
scikit-learn==1.5.2
optuna==4.1.0
shap==0.46.0

# API
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.28.0

# Scheduling
apscheduler==3.10.4
